pyproject.toml
[project]
name = "xelbot"
version = "0.1.0"
description = "Discord bot"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "discord-py>=2.5.2",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "trafilatura>=2.0.0",
]

[project.scripts]
xelbot = "xelbot.cli:cli"

[build-system]
requires = ["uv_build >= 0.7.19, <0.9.0"]
build-backend = "uv_build"